task-semifeita

+++ Core/NeuroGraph.Main/Entities/BiologicalSignal.cs
@@ -5,7 +5,7 @@ namespace Api.Entities
        public Guid Id { get; set; }
         public Guid NeuronId { get; set; }          // FK
         public Neuron Neuron { get; set; } = null!;
-        public double Intensity { get; set; }
+        public float Intensity { get; set; }
         public DateTime FiredAt { get; set; } = DateTime.UtcNow;
     }
 }
